Create adapter to execute use case
It was necessary to create an adapter for the use case, otherwise it would be necessary to use the @injectable decorator in the original use case, which would violate the principles of clean architecture. Use case is in the domain layer. Injectable decorater is a resource of Nest JS, which is a fframework, which belongs to the most external layer, i.e., Frameworks & Drivers layer